Just call populate once, and let the tests do the querying.
